packages/asp-core/pyproject.toml
[project]
name = "asp-core"
version = "0.1.0"
description = "Agentic Security Platform — core domain (pure, no I/O)"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
dependencies = [
    "pydantic>=2.9",
    "pyyaml>=6.0",
    "structlog>=24.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/asp_core"]

# Ontology YAML files ship with the wheel.
[tool.hatch.build.targets.wheel.force-include]
"src/asp_core/graph/ontology" = "asp_core/graph/ontology"

[tool.mypy]
# Core must be strict. No exceptions.
python_version = "3.12"
strict = true
warn_unreachable = true
disallow_any_explicit = true
